+++ CHANGELOG.md
@@ -6,3 +6,14 @@
 - remove type `ISpecifier`
 - remove helper `moduleExport`
 - change `addAppExport(string, ISpecifier)` to `addAppExport(string, string)`
+
+
+### router
+
+https://docs.google.com/document/d/1g3KPpEdl9DILftI22qWKUQIVbAxtkoGKSTbEZAFjxDg/edit#
+
+
+### app
+
+- remove `parsedUrl` from `req` 
+- add `query` and `pathname` to `req`